chore: move explanation, multi_explanation, evaluation, and rex to new explanation submodule
