+++ rex_xai/output/database.py
@@ -13,8 +13,8 @@ import numpy as np
 
 from rex_xai.utils.logger import logger
 from rex_xai.input.config import CausalArgs, Strategy
-from rex_xai.explanation import Explanation
-from rex_xai.multi_explanation import MultiExplanation
+from rex_xai.explanation.explanation import Explanation
+from rex_xai.explanation.multi_explanation import MultiExplanation
 
 
 def _dataframe(db, table):
