+++ tests/conftest.py
@@ -6,15 +6,15 @@ from rex_xai.utils._utils import get_device
 from rex_xai.mutants.box import initialise_tree
 from rex_xai.input.config import CausalArgs, process_custom_script, Strategy
 from rex_xai.mutants.distributions import Distribution
-from rex_xai.rex import (
+from rex_xai.explanation.rex import (
     calculate_responsibility,
     get_prediction_func_from_args,
     load_and_preprocess_data,
     predict_target,
     try_preprocess,
 )
-from rex_xai.explanation import Explanation
-from rex_xai.multi_explanation import MultiExplanation
+from rex_xai.explanation.explanation import Explanation
+from rex_xai.explanation.multi_explanation import MultiExplanation
 from syrupy.extensions.amber.serializer import AmberDataSerializer
 from syrupy.filters import props
 from syrupy.matchers import path_type
