+++ tests/unit_tests/data_test.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import numpy as np
-from rex_xai.rex import load_and_preprocess_data
+from rex_xai.explanation.rex import load_and_preprocess_data
 from rex_xai.input.input_data import Data
 
 tab = np.arange(0, 1, 999)
