+++ tests/unit_tests/multi-explanation_test.py
@@ -4,9 +4,9 @@ import os
 import numpy as np
 import pytest
 from rex_xai.utils._utils import Strategy
-from rex_xai.rex import calculate_responsibility
-from rex_xai.explanation import Explanation
-from rex_xai.multi_explanation import MultiExplanation
+from rex_xai.explanation.rex import calculate_responsibility
+from rex_xai.explanation.explanation import Explanation
+from rex_xai.explanation.multi_explanation import MultiExplanation
 
 
 @pytest.mark.parametrize("spotlights", [5, 10])
